panagia: implement Single-Decree Paxos
Based on the TLA+ spec of Single-Decree Paxos as found at
https://raw.githubusercontent.com/tlaplus/tlapm/581ebd8afbe927b2b7d31c62ccec3a3409d9d8ce/examples/paxos/Paxos.tla
this commit brings a (model-only) implementation of the protocol,
using MTL-style classes defining the effects various actors in the
protocol need.

The intent is for the implementation to closely follow the TLA+ spec
(hence the spec code is provided as comments), and for it to fit on a
single piece of paper, so to say.

Next to this, the `Panagia.Paxos.SingleDecree.Free` module exposes
functors which can be used to implement these classes using a "free
monad" approach (as, e.g., is done in the test-suite).

Finally, the implementation is validated using various tests, some as
unit-tests, some using Hedgehog to emulate message reordering, message
duplication and message loss. The latter did find a bug in an earlier
implementation of the protocol during development!

Closes: https://github.com/NicolasT/panagia/issues/4
Closes: https://github.com/NicolasT/panagia/issues/5