panagia: implement `Learner` functionality
Similar to how `Proposer` and `Accepted` are implemented, this provides
an MTL-style class of effects, and a handler, in this case for
`Accepted` messages.

This handler takes, similar to `propose`, a `LearnerState`, which *must*
be constructed by the caller when the learner starts (using
`initLearnerState`). An invocation of `handleAccepted` may then either
yield a value, if consensus was reached (and the state can be
discarded), or an updated state to be used in a next call to
`handleAccepted`.

Test-cases are updated accordingly.

Closes: https://github.com/NicolasT/panagia/issues/15